+++ lib/cocoapods/sources_manager.rb
@@ -109,7 +109,7 @@ module Pod
         if source_name
           sources = [git_source_named(source_name)]
         else
-          sources =  git_sources
+          sources = git_sources
         end
 
         changed_spec_paths = {}
@@ -127,6 +127,14 @@ module Pod
         end
         # Perform search index update operation in background.
         update_search_index_if_needed_in_background(changed_spec_paths)
+      end
+
+      # Adds the provided source to the list of sources
+      #
+      # @param [Source] source the source to add
+      #
+      def add_source(source)
+        all << source unless all.any? { |s| s.url == source || s.name == source.name }
       end
     end